tests/wpt/tests/css/css-grid/placement/grid-auto-placement-implicit-tracks-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Grid items auto-placement implicit tracks</title>
<link rel="author" title="Manuel Rego Casasnovas" href="mailto:rego@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-grid-1/#auto-placement-algo">
<meta name="assert" content="This test checks that auto-placement works properly even when implicit tracks have been created during the first steps of the placement algorithm.">
<style>
.grid {
    display: grid;
    position: relative;
    grid-auto-columns: 100px;
    grid-auto-rows: 50px;
    margin: 10px;
}
.gridAutoFlowColumn { grid-auto-flow: column; }
.gridAutoFlowColumnDense { grid-auto-flow: column dense; }
.gridAutoFlowRowDense { grid-auto-flow: row dense; }
.grid > div { border: solid; }
</style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>

<body onload="checkLayout('.grid')">
<div id="log"></div>

<div class="grid">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: span 2;"
      data-offset-x="0" data-offset-y="50" data-expected-width="200" data-expected-height="100">i2</div>
  <div data-offset-x="0" data-offset-y="150" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="100" data-offset-y="150" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: 1; grid-row: span 2;"
      data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="100">i2</div>
  <div data-offset-x="0" data-offset-y="150" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="0" data-offset-y="200" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid">
  <div data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: 1;"
      data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="50">i2</div>
  <div data-offset-x="100" data-offset-y="50" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="0" data-offset-y="100" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowRowDense">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: span 2;"
      data-offset-x="0" data-offset-y="50" data-expected-width="200" data-expected-height="100">i2</div>
  <div data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="0" data-offset-y="150" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowRowDense">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: 1; grid-row: span 2;"
      data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="100">i2</div>
  <div data-offset-x="0" data-offset-y="150" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="0" data-offset-y="200" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowRowDense">
  <div data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: 1;"
      data-offset-x="0" data-offset-y="0" data-expected-width="200" data-expected-height="50">i2</div>
  <div data-offset-x="100" data-offset-y="50" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="0" data-offset-y="100" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumn">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: span 2;"
      data-offset-x="100" data-offset-y="0" data-expected-width="200" data-expected-height="100">i2</div>
  <div data-offset-x="300" data-offset-y="0" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="300" data-offset-y="50" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumn">
  <div data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: 1; grid-row: span 2;"
      data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="100">i2</div>
  <div data-offset-x="100" data-offset-y="50" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumn">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: 1;"
      data-offset-x="100" data-offset-y="0" data-expected-width="200" data-expected-height="50">i2</div>
  <div data-offset-x="300" data-offset-y="0" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="400" data-offset-y="0" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumnDense">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: span 2;"
      data-offset-x="100" data-offset-y="0" data-expected-width="200" data-expected-height="100">i2</div>
  <div data-offset-x="0" data-offset-y="50" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="300" data-offset-y="0" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumnDense">
  <div data-offset-x="100" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: 1; grid-row: span 2;"
      data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="100">i2</div>
  <div data-offset-x="100" data-offset-y="50" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="200" data-offset-y="0" data-expected-width="100" data-expected-height="50">i4</div>
</div>

<div class="grid gridAutoFlowColumnDense">
  <div data-offset-x="0" data-offset-y="0" data-expected-width="100" data-expected-height="50">i1</div>
  <div style="grid-column: span 2; grid-row: 1;"
      data-offset-x="100" data-offset-y="0" data-expected-width="200" data-expected-height="50">i2</div>
  <div data-offset-x="300" data-offset-y="0" data-expected-width="100" data-expected-height="50">i3</div>
  <div data-offset-x="400" data-offset-y="0" data-expected-width="100" data-expected-height="50">i4</div>
</div>

</body>
